// src/lib/api/generated/model/profileResDto.ts
/**
 * Generated by orval v8.14.0 🍺
 * Do not edit manually.
 * BE Global Service API
 * BE Global Service API
 * OpenAPI spec version: 1.0.0
 */
import type { CountryResDto } from './countryResDto';
import type { InstantContactDto } from './instantContactDto';
import type { ProfileResDtoCurrency } from './profileResDtoCurrency';
import type { ProfileResDtoLanguage } from './profileResDtoLanguage';
import type { ProfileResDtoRole } from './profileResDtoRole';
import type { ProfileResDtoUserType } from './profileResDtoUserType';
import type { StateResDto } from './stateResDto';

export interface ProfileResDto {
  id: string;
  avatar?: string;
  email: string;
  phoneCountry?: CountryResDto;
  dialling: string;
  phoneNumber: string;
  fullName: string;
  representative: string;
  taxId: string;
  postalCode: string;
  city: string;
  address: string;
  instantContacts?: InstantContactDto[];
  userType: ProfileResDtoUserType;
  language: ProfileResDtoLanguage;
  currency: ProfileResDtoCurrency;
  country?: CountryResDto;
  state?: StateResDto;
  role: ProfileResDtoRole;
}